add deepequal for pkg/policy
Not relying on reflect.DeepEquals improves the DeepEqual performance
logic a lot. As an example, a benchmark was written for SpecEquals and
this commit was able to reduce the number of allocations and bytes per
operation to 0.

```
BenchmarkSpecEquals/Reflected_SpecEquals-8     98300  12325 ns/op  3117 B/op  68 allocs/op
BenchmarkSpecEquals/Generated_SpecEquals-8   2628786    452 ns/op     0 B/op   0 allocs/op
```

Signed-off-by: André Martins <andre@cilium.io>