Return error when deleting ENI rules fails
[ upstream commit 820d852583fc736131de8cecbbe0632b77a4ea88 ]

Previously, we relied on a log message to indicate whether a deletion of
a rule failed. Relying on the log message can be potentially misleading
as endpoint deletion is best-effort and errors are ignored in the end.
The user may see log message indicating a rule was deleted, but later in
the endpoint deletion it would show as an error.

This commit changes the function which deletes rules to surface the
error. This change removes the misleading log messages and allows us to
assert against this in tests.

Signed-off-by: Chris Tarazi <chris@isovalent.com>